Fix tests
After merging master into my branch, when I attempted to run the
tests I got this error: "OperationFailedError: No active warehouse
selected in the current session.  Select an active warehouse with
the 'use warehouse' command."

It looks like I was relying on the "last used warehouse" before.
Now I've changed the tests to use the configured warehouse for each
connection (this is configured with the SNOWFLAKE_TEST_WAREHOUSE
environment variable).